+++ demo/taxonomy-browser/taxonomy-browser.js
@@ -28,21 +28,11 @@ var taxonTree, searchResultTree, tmplDetails, tmplInfoPane, tmplMedia,
 	SEARCH_PAGE_SIZE = 5,
 	CHILD_NODE_PAGE_SIZE = 200,
 	glyphOpts = {
+		preset: "bootstrap3",
 		map: {
-			doc: "glyphicon glyphicon-file",
-			docOpen: "glyphicon glyphicon-file",
-			checkbox: "glyphicon glyphicon-unchecked",
-			checkboxSelected: "glyphicon glyphicon-check",
-			checkboxUnknown: "glyphicon glyphicon-share",
-			dragHelper: "glyphicon glyphicon-play",
-			dropMarker: "glyphicon glyphicon-arrow-right",
-			error: "glyphicon glyphicon-warning-sign",
-			expanderClosed: "glyphicon glyphicon-menu-right",
+			expanderClosed: "glyphicon glyphicon-menu-right",  // glyphicon-plus-sign
 			expanderLazy: "glyphicon glyphicon-menu-right",  // glyphicon-plus-sign
-			expanderOpen: "glyphicon glyphicon-menu-down",  // glyphicon-collapse-down
-			folder: "glyphicon glyphicon-folder-close",
-			folderOpen: "glyphicon glyphicon-folder-open",
-			loading: "glyphicon glyphicon-refresh glyphicon-spin"
+			expanderOpen: "glyphicon glyphicon-menu-down"  // glyphicon-collapse-down
 		}
 	};