+++ src/jquery.fancytree.dnd5.js
@@ -34,7 +34,7 @@
 /* *****************************************************************************
  * Private functions and variables
  */
-var
+var FT = $.ui.fancytree,
 	classDragSource = "fancytree-drag-source",
 	classDragRemove = "fancytree-drag-remove",
 	classDropAccept = "fancytree-drop-accept",
@@ -232,7 +232,7 @@ function handleDragOver(event, data) {
 			.toggleClass(classDropOver, hitMode === "over")
 			.toggleClass(classDropBefore, hitMode === "before")
 			.show()
-			.position($.ui.fancytree.fixPositionOptions({
+			.position(FT.fixPositionOptions({
 				my: "left" + offsetString(markerOffsetX) + " center",
 				at: "left " + markerAt,
 				of: $targetTitle
@@ -291,7 +291,7 @@ $.ui.fancytree.registerExtension({
 			opts = ctx.options,
 			glyph = opts.glyph || null,
 			dndOpts = opts.dnd5,
-			getNode = $.ui.fancytree.getNode;
+			getNode = FT.getNode;
 
 		if( $.inArray("dnd", opts.extensions) >= 0 ) {
 			$.error("Extensions 'dnd' and 'dnd5' are mutually exclusive.");
@@ -303,7 +303,7 @@ $.ui.fancytree.registerExtension({
 		// Implement `opts.createNode` event to add the 'draggable' attribute
 		// #680: this must happen before calling super.treeInit()
 		if( dndOpts.dragStart ) {
-			$.ui.fancytree.overrideMethod(ctx.options, "createNode", function(event, data) {
+			FT.overrideMethod(ctx.options, "createNode", function(event, data) {
 				// Default processing if any
 				this._super.apply(this, arguments);
 
@@ -328,7 +328,8 @@ $.ui.fancytree.registerExtension({
 					"pointer-events": "none"
 				}).prependTo("body");
 			if( glyph ) {
-				$dropMarker.addClass(glyph.map.dropMarker);
+				FT.setSpanIcon($dropMarker[0], glyph.map._addClass, glyph.map.dropMarker);
+				// $dropMarker.addClass(glyph.map._addClass + " " + glyph.map.dropMarker);
 			}
 		}
 		// Enable drag support if dragStart() is specified: